backend/requirements.txt
# SkillSync GraphRAG Backend - Dependencies
# Generated from pip freeze on 2026-01-29

# Core Framework
fastapi==0.128.0
uvicorn[standard]==0.40.0

# Database
motor==3.7.1

# Graph Processing
networkx==3.5

# Data Validation
pydantic==2.12.5
pydantic-settings>=2.0.0
email-validator>=2.1.0

# Environment
python-dotenv==1.2.1

# HTTP Client (for tests)
httpx==0.27.0

# Development
ruff>=0.14.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
